+++ lib/ts/recipe/thirdpartyemailpassword/types.ts
@@ -17,6 +17,7 @@ import { TypeInput as TypeInputEmailVerification } from "../emailverification/ty
 import {
     RecipeInterface as EmailVerificationRecipeInterface,
     APIInterface as EmailVerificationAPIInterface,
+    OriginalAPIInterface as EmailVerificationOriginalAPIInterface,
 } from "../emailverification";
 import {
     NormalisedFormField,
@@ -156,10 +157,10 @@ export type TypeInput = {
     emailVerificationFeature?: TypeInputEmailVerificationFeature;
     override?: {
         functions?: (originalImplementation: RecipeInterface) => RecipeInterface;
-        apis?: (originalImplementation: APIInterface) => APIInterface;
+        apis?: (originalImplementation: OriginalAPIInterface) => APIInterface;
         emailVerificationFeature?: {
             functions?: (originalImplementation: EmailVerificationRecipeInterface) => EmailVerificationRecipeInterface;
-            apis?: (originalImplementation: EmailVerificationAPIInterface) => EmailVerificationAPIInterface;
+            apis?: (originalImplementation: EmailVerificationOriginalAPIInterface) => EmailVerificationAPIInterface;
         };
     };
 };
@@ -185,10 +186,10 @@ export type TypeNormalisedInput = {
     emailVerificationFeature: TypeInputEmailVerification;
     override: {
         functions: (originalImplementation: RecipeInterface) => RecipeInterface;
-        apis: (originalImplementation: APIInterface) => APIInterface;
+        apis: (originalImplementation: OriginalAPIInterface) => APIInterface;
         emailVerificationFeature?: {
             functions?: (originalImplementation: EmailVerificationRecipeInterface) => EmailVerificationRecipeInterface;
-            apis?: (originalImplementation: EmailVerificationAPIInterface) => EmailVerificationAPIInterface;
+            apis?: (originalImplementation: EmailVerificationOriginalAPIInterface) => EmailVerificationAPIInterface;
         };
     };
 };
@@ -308,6 +309,47 @@ export type SignInUpAPIOutput =
           status: "FIELD_ERROR";
           error: string;
       };
+
+export interface OriginalAPIInterface {
+    authorisationUrlGET: (input: {
+        provider: TypeProvider;
+        options: ThirdPartyAPIOptions;
+    }) => Promise<{
+        status: "OK";
+        url: string;
+    }>;
+
+    emailExistsGET: (input: {
+        email: string;
+        options: EmailPasswordAPIOptions;
+    }) => Promise<{
+        status: "OK";
+        exists: boolean;
+    }>;
+
+    generatePasswordResetTokenPOST: (input: {
+        formFields: {
+            id: string;
+            value: string;
+        }[];
+        options: EmailPasswordAPIOptions;
+    }) => Promise<{
+        status: "OK";
+    }>;
+
+    passwordResetPOST: (input: {
+        formFields: {
+            id: string;
+            value: string;
+        }[];
+        token: string;
+        options: EmailPasswordAPIOptions;
+    }) => Promise<{
+        status: "OK" | "RESET_PASSWORD_INVALID_TOKEN_ERROR";
+    }>;
+
+    signInUpPOST: (input: SignInUpAPIInput) => Promise<SignInUpAPIOutput>;
+}
 
 export interface APIInterface {
     authorisationUrlGET: