+++ CHANGELOG.md
@@ -5,9 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [6.2.0] - 2021-07-21
+## [7.0.0] - 2021-07-21
 
-## Changes
+## Breaking change:
 
 -   Updated minimum Oauth scopes required for various thirdparty providers.
 -   For thirdparty providers, if the scope array is passed by the user, the default scopes will not be used.